[FIX] l10n_it_edi_sdicoop: relocate relevant tests
Tests that were added to l10n_it_edi for price included taxes and
discounted invoices a0a8bdf should actually belong in
l10n_it_edi_sdicoop (partially because of the requirement for the proxy
user when l10n_it_edi_sdicoop is installed). This disctinction will
eventually be irrelevant in master as the two modules should be merged
together eventually.

Some data was added on the Italian company in the python and the XML.
This is because when l10n_it_edi_sdicoop is installed (as opposed to
just l10n_it_edi), conditionals have to be met before posting an
invoice, (e.g. the company must have a street, city, postcode, etc).
