+++ addons/l10n_it_edi_sdicoop/tests/expected_xmls/IT00470550013_basis.xml
@@ -21,10 +21,13 @@
         <Anagrafica>
           <Denominazione>company_2_data</Denominazione>
         </Anagrafica>
+        <RegimeFiscale>RF01</RegimeFiscale>
       </DatiAnagrafici>
       <Sede>
-        <Indirizzo> </Indirizzo>
-        <CAP>00000</CAP>
+        <Indirizzo>1234 Test Street </Indirizzo>
+        <CAP>12345</CAP>
+        <Comune>Prova</Comune>
+        <Nazione>IT</Nazione>
       </Sede>
     </CedentePrestatore>
     <CessionarioCommittente>
@@ -41,6 +44,7 @@
       <Sede>
         <Indirizzo>Via Privata Alessi 6 </Indirizzo>
         <CAP>28887</CAP>
+        <Comune>Milan</Comune>
         <Nazione>IT</Nazione>
       </Sede>
     </CessionarioCommittente>
@@ -59,6 +63,7 @@
     <DatiPagamento>
       <CondizioniPagamento>TP02</CondizioniPagamento>
       <DettaglioPagamento>
+        <ModalitaPagamento>MP05</ModalitaPagamento>
         <DataScadenzaPagamento>2022-03-24</DataScadenzaPagamento>
         <ImportoPagamento></ImportoPagamento>
         <CodicePagamento>___ignore___</CodicePagamento>
